Close memcache connections in example-ruby-memcache
example-ruby-memcache doesn't close connections to the
Memcache server (or Memcache SP); every request to the
app creates a client and uses that client to make a
request and return the server's response. But the
connection is not closed explicitly.

The Memcache SP expects EOF when the client disconnects;
EOF causes the SP stop reading from the connection and
exit. If EOF is not found at the end of the connection,
the reads made by the SP will timeout and it will retry
again indefinetely.

Have each request to example-ruby-memcache close its
Memcache connection, so that the SP read EOF and exit.

Fixes ENGT-12807